LayoutTests/platform/glib/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-import-data-url.any.worker-expected.txt

CONSOLE MESSAGE: Cross-origin redirection to http://web-platform.test:8800/workers/modules/resources/export-on-load-script.js denied by Cross-Origin Resource Sharing policy: Origin null is not allowed by Access-Control-Allow-Origin. Status code: 302

Harness Error (FAIL), message = Error in remote: Importing a module script failed.

PASS Static import.
PASS Static import (cross-origin).
TIMEOUT Static import (redirect). Test timed out
NOTRUN Nested static import.
NOTRUN Static import and then dynamic import.
NOTRUN Dynamic import.
NOTRUN Nested dynamic import.
NOTRUN Dynamic import and then static import.
NOTRUN eval(import()).
